fix(agents): release expired prompt reservations
